int-tests: suggestions: rename the withapim setup classes
